+++ python-client-sdk/src/sdk/sdk.py
@@ -22,7 +22,7 @@ class SDK:
     _server_url: str = SERVERS[0]
     _language: str = "python"
     _sdk_version: str = "1.1.1"
-    _gen_version: str = "2.23.4"
+    _gen_version: str = "2.23.6"
 
     def __init__(self,
                  server_url: str = None,