+++ RELEASES.md
@@ -64,3 +64,12 @@ Based on:
 ### Releases
 - [PyPI v1.1.1] https://pypi.org/project/brease-sdk/1.1.1 - python-client-sdk
 - [NPM v0.1.1] https://www.npmjs.com/package/brease-sdk/v/0.1.1 - typescript-client-sdk
+
+## 2023-05-02 00:41:03
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 https://prod.speakeasyapi.dev/v1/apis/brease/version/0.1/schema/download
+- Speakeasy CLI 1.26.5 (2.23.6) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [PyPI v1.1.1] https://pypi.org/project/brease-sdk/1.1.1 - python-client-sdk
+- [NPM v0.1.1] https://www.npmjs.com/package/brease-sdk/v/0.1.1 - typescript-client-sdk